Table: keyed rows
prevent occasional corruption of vnode.state/this upon deletion of rows. When tables hold components in each row, sometimes the dom node(s) for the wrong rows are deleted. When this happens, a mixture of invalidated and offset component state is paired up with the redraw attrs, leading to desynchronization in all subsequent redraws until the component is reinitialized